# Faster decompression of non-fdeflate encoded PNGs