@fintelia 2023-03-01
Interestingly the results are reversed if I first re-encode each image with fdeflate (making the zlib stream incredibly literal heavy):
```
Directory                                     Ratio             zune-png                  Decode
---------                                    -------     --------------------      --------------------
Total                                         0.000%     150 mps  0.509 GiB/s      183 mps  0.622 GiB/s
```